improve condor submission

+++ sf_inclusion/condor/eff_and_sf.py
@@ -62,7 +62,7 @@ def eff_and_sf(args):
     jw.write_condor( filename=outs_submit,
                      executable=outs_job,
                      outfile=outs_check,
-                     queue='short',
+                     queue='long',
                      machine='llrt3condor' )
 
     qlines = []
