+++ sf_inclusion/condor/eff_and_sf_aggr.py
@@ -48,7 +48,7 @@ def eff_and_sf_aggr(args):
     jw.write_condor( filename=outs_submit,
                      executable=outs_job,
                      outfile=outs_check,
-                     queue='short',
+                     queue='long',
                      machine='llrt3condor' )
 
     qlines = []
