+++ sf_inclusion/condor/hadd_eff.py
@@ -83,7 +83,7 @@ def hadd_eff(args):
         jw.write_condor( filename=out2,
                          executable=out1,
                          outfile=out3,
-                         queue='short',
+                         queue='long',
                          machine='llrt3condor' )
 
         qlines = []
